review: consolidate tests into transpiler.test.js, add types and docs
- Move 12 new tests from transpiler-minify-dce.test.ts into transpiler.test.js
  per CLAUDE.md convention (add to existing file, don't create new)
- Add minifySyntax and minifyIdentifiers to TranspilerOptions in bun-types
- Document new options in docs/runtime/transpiler.mdx
