+++ docs/runtime/transpiler.mdx
@@ -234,6 +234,12 @@ interface TranspilerOptions {
   // Experimental whitespace minification
   minifyWhitespace?: boolean,
 
+  // Experimental syntax minification (constant folding, dead code elimination)
+  minifySyntax?: boolean,
+
+  // Experimental identifier minification (rename to shorter names)
+  minifyIdentifiers?: boolean,
+
   // Whether to inline constant values
   // Typically improves performance and decreases bundle size
   // Default: true
