+++ packages/bun-types/bun.d.ts
@@ -2470,6 +2470,19 @@ declare module "bun" {
     /**
      * **Experimental**
      *
+     * Fold constant expressions, eliminate dead branches, and perform other
+     * syntax-level minifications.
+     */
+    minifySyntax?: boolean;
+    /**
+     * **Experimental**
+     *
+     * Rename local identifiers to shorter names.
+     */
+    minifyIdentifiers?: boolean;
+    /**
+     * **Experimental**
+     *
      * Enabled by default, use this to disable dead code elimination.
      *
      * Some other transpiler options may still do some specific dead code elimination.
